fun-fact-mcp-server/pyproject.toml
[project]
name = "fun-fact-mcp-server"
version = "0.1.0"
description = "An Example MCP server on Cloud Run"
readme = "README.md"
requires-python = ">=3.12"
dependencies = ["fastmcp==2.12.3", "google-cloud-compute==1.37.0"]
